Adds wiring for gitpod and vscode.
Overall:

1. Moves to make goals both clearer and broader to allow for more
research and broader application.
2. Adds wiring for gitpod to be able to create a consistent development
environment. The development environment is presently not minimal and
is more focused on getting a setup running smoothly and consistently.
3. Adds wiring for VS Code and a basic, common set of plugins to make
development easy.
4. Corrects a buch of linter errors, reformats things to fit with the
(now enforced) standards, etc.

None of this is to say that anyone _has_ to develop a particular way and
future work can definitely trim this down or focus on a more consistent,
minimal set. For the moment the goal is just to be able to bootstrap any
developer who wants to contribute with a sensible set of customizations.